+++ inst/htmlwidgets/lib/kml/kml.js
@@ -1,16 +1,18 @@
-function add_kml(map_id, kml_data, layer_id) {
+function add_kml(map_id, kml_data, layer_id, update_map_view) {
 
     //window[map_id + 'googleKml' + layer_id] = [];
     createWindowObject(map_id, 'googleKml', layer_id);
 
-
     var kmlLayer = new google.maps.KmlLayer({
         url: kml_data[0].url,
         map: window[map_id + 'map'],
+        preserveViewport: update_map_view,
         zIndex: kml_data[0].z_index
     });
 
-    window[map_id + 'googleKml' + layer_id].push(kmlLayer);
+  console.log(kmlLayer);
+
+  window[map_id + 'googleKml' + layer_id].push(kmlLayer);
 }