+++ inst/htmlwidgets/lib/overlay/overlay.js
@@ -1,6 +1,6 @@
 function add_overlay (map_id, data_overlay, layer_id) {
 
-  //window[map_id + 'googleOverlay' + layer_id] = [];
+    console.log("adding overlay");
     createWindowObject(map_id, 'googleOverlay', layer_id);
 
     var latlonNorthEast = new google.maps.LatLng(data_overlay[0].north, data_overlay[0].east);
@@ -23,17 +23,20 @@ function add_overlay (map_id, data_overlay, layer_id) {
     window[map_id + 'mapBounds'].extend(latlonNorthEast);
     window[map_id + 'mapBounds'].extend(latlonSouthWest);
 
-    window[map_id + 'map'].fitBounds(window[map_id + 'mapBounds'])
+    window[map_id + 'map'].fitBounds(window[map_id + 'mapBounds']);
     window[map_id + 'googleOverlay' + layer_id].push(overlayLayer);
 
 }
 
 
 function clear_overlay (map_id, layer_id) {
-    
+
+  //console.log(window[map_id + 'googleOverlay' + layer_id]);
+
 //    for (i = 0; i < window[map_id + 'googleOverlay' + layer_id].length; i++){
 //        window[map_id + 'googleOverlay' + layer_id][i].setMap(null);
 //    }
 //    window[map_id + 'googleOverlay' + layer_id] = null;
+//    window[map_id + 'googleOverlay' + layer_id].setMap(null);
     clear_object(map_id, 'googleOverlay', layer_id);
 }